Global refactoring and adding new functionality
docScript was removed from VkScriptCore and moved to separate files.
Added the function of saving user information to a file together with the program results.
User info in the docScript is the user name, user id and the scanned interval(beta).
Now the 'tokens' file stores a lot of tokens and offers a choice for authentication.
More logs.
And probably something else.